# docs/swagger.yaml
basePath: /api
definitions:
  ApiResponse:
    description: Generic API response
    properties:
      data: {}
      responseKey:
        type: string
    type: object
...
  ServicesCount:
    description: ServicesCount represents the count of services.
    properties:
      count:
        description: Count is the number of services.
        type: integer
    type: object
  ServicesDetails:
    description: ServicesDetails represents detailed information about a specific
      service.
    properties:
      createdAt:
        description: CreatedAt is the timestamp indicating when the service entry
          was created.
        type: string
      description:
        description: description is the description of the service. It is required
          and should be between 3 and 35 characters.
        maxLength: 35
        minLength: 3
        type: string
      id:
        description: ID is the unique identifier for the user.
        type: string
      name:
        description: name is the  name of the service. It is required and should be
          between 3 and 30 characters.
        maxLength: 30
        minLength: 3
        type: string
    required:
    - description
    - name
    type: object
  ServicesIn:
    description: ServicesIn represents the input structure for creating a new service.
    properties:
      companyID:
        description: CompanyID is the unique identifier for the company associated
          with the service. It is required.
        type: string
      description:
        description: description is the description of the service. It is required
          and should be between 3 and 35 characters.
        maxLength: 35
        minLength: 3
        type: string
      keywordName:
        description: KeywordName is the name of the keyword's service. It is required
          and should be between 2 and 255 characters.
        maxLength: 255
        minLength: 2
        type: string
      name:
        description: name is the  name of the service. It is required and should be
          between 3 and 30 characters.
        maxLength: 30
        minLength: 3
        type: string
    required:
    - companyID
    - description
    - keywordName
    - name
    type: object
  ServicesList:
    description: ServicesList represents a simplified version of the service for listing
      purposes.
    properties:
      id:
        description: ID is the unique identifier for the service.
        type: string
      name:
        description: Name is the full name of the service.
        type: string
    type: object
  ServicesPagination:
    description: ServicesPagination represents the paginated list of Services.
    properties:
      items:
        description: Items is a slice containing individual service details.
        items:
          $ref: '#/definitions/ServicesTable'
        type: array
      limit:
        description: Limit is the maximum number of items per page in the pagination.
        type: integer
      page:
        description: Page is the current page number in the pagination.
        type: integer
      totalCount:
        description: TotalCount is the total number of services in the entire list.
        type: integer
    type: object
  ServicesTable:
    description: KeywordsTable represents a single service entry in a table.
    properties:
      createdAt:
        description: CreatedAt is the timestamp indicating when the service entry
          was created.
        type: string
      id:
        description: ID is the unique identifier for the user.
        type: string
      name:
        description: name is the  name of the service. It is required and should be
          between 3 and 30 characters.
        maxLength: 30
        minLength: 3
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: h.manai@tunisiancloud.com
    name: Hedi Manai
  description: A basic template for managing your project using Golang REST APIs.
  title: Project Name
  version: "0.1"
paths:
...
  /services/{companyID}:
    get:
      description: Get all services.
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ServicesPagination'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get services
      tags:
      - Services
    post:
      consumes:
      - application/json
      description: Create a new service.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: Service query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ServicesIn'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Create service
      tags:
      - Services
  /services/{companyID}/{ID}:
    delete:
      description: Delete one service.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: service ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete service
      tags:
      - Services
    get:
      description: Get one service.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: Service ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ServicesDetails'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get service
      tags:
      - Services
    put:
      consumes:
      - application/json
      description: Update service.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: Service ID
        in: path
        name: ID
        required: true
        type: string
      - description: Service query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ServicesIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Update service
      tags:
      - Services
  /services/{companyID}/count:
    get:
      description: Get number of all services.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ServicesCount'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get number of  services
      tags:
      - Services
  /services/{companyID}/list:
    get:
      description: Get list of all services.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ServicesList'
            type: array
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get list of  services
      tags:
      - Services
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"